requirements_dev.txt
pip==9.0.1
bumpversion==0.5.3
wheel==0.29.0
watchdog==0.8.3
flake8==3.2.1
tox==2.3.1
coverage==4.1
Sphinx==1.4.8
sphinx-autobuild
docutils<0.13.1
cryptography==1.4
PyYAML==3.11
pytest==2.9.2
ipython==5.1.0
pysqlite==2.8.3
